# Update performance characteristics to show initial, instant, and pending build times