This PR updates the performance characteristics section to display 3 "build time" metrics. They are:

- Initial build time: The build time of the first build
- Instant build time savings: The immediate savings as a result of build caching. 
- Pending build time savings: An estimate of the potential savings if all executed cacheable tasks had come from cache.

Example output of Gradle experiments 2-5 and Maven experiments 1-4:

```
Performance Characteristics
---------------------------
Initial build time:              16.284s
Build time with instant savings: 3.520s, 12.764s savings
Build time with pending savings: 3.520s, 0.000s additional savings
Avoided cacheable tasks:          63 tasks, 1m 21.524s total saved execution time
Executed cacheable tasks:          0 tasks, 0.000s total execution time
Executed non-cacheable tasks:    118 tasks, 6.167s total execution time
Serialization factor:            4.97x
```

Example output of Gradle experiment 1:

```
Performance Characteristics
---------------------------
Initial build time:              23.605s
Build time with instant savings: 1.945s, 21.660s savings
Serialization factor:            3.59x
```